+++ src/main/java/com/lj/exam/model/ExamRecord.java
@@ -28,9 +28,9 @@ public class ExamRecord {
 		this.examSiteSeat = examSiteSeat;
 	}
 	
-	private Long videoScore;
+	private Double videoScore;
 	
-	private Long fileScore;
+	private Double fileScore;
 
 	private String status;
 
@@ -102,19 +102,19 @@ public class ExamRecord {
 		this.examineeName = examineeName;
 	}
 
-	public Long getVideoScore() {
+	public Double getVideoScore() {
 		return videoScore;
 	}
 
-	public void setVideoScore(Long videoScore) {
+	public void setVideoScore(Double videoScore) {
 		this.videoScore = videoScore;
 	}
 
-	public Long getFileScore() {
+	public Double getFileScore() {
 		return fileScore;
 	}
 
-	public void setFileScore(Long fileScore) {
+	public void setFileScore(Double fileScore) {
 		this.fileScore = fileScore;
 	}
